# openapi/profiles/refund-assistant.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SupportPilot Actions — refund-assistant",
    "description": "Business operations the SupportPilot agent may request on behalf of an authenticated support user. Every operation authorizes the individual user; none accepts an identity supplied by the caller.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://broker:8097/refund-assistant",
      "description": "The delegation broker, as the refund-assistant profile. Every call is made with a token naming the user and refund-assistant, narrowed to the call."
    }
  ],
  "paths": {
    "/v1/orders/{order_number}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get an order visible to the authenticated user",
        "operationId": "get_order",
        "parameters": [
          {
            "name": "order_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 16,
              "pattern": "^ORD-[0-9]{4,12}$",
              "description": "Order number in the form ORD-nnnn",
              "examples": [
                "ORD-2001"
              ],
              "title": "Order Number"
            },
            "description": "Order number in the form ORD-nnnn"
          },
          {
            "name": "include_items",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include the order's line items",
              "default": false,
              "title": "Include Items"
            },
            "description": "Include the order's line items"
          },
          {
            "name": "include_shipment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include shipment and tracking status",
              "default": false,
              "title": "Include Shipment"
            },
            "description": "Include shipment and tracking status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Not found or not allowed"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
    "/v1/actions/refunds": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Propose a refund for review. This does not move money.",
        "description": "Creates a pending refund request that an independent approver must review. No money moves when this succeeds. Tell the user their request was submitted for approval — never that the refund has been issued.",
        "operationId": "propose_refund",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "order_number": {
                    "type": "string",
                    "maxLength": 16,
                    "minLength": 8,
                    "pattern": "^ORD-[0-9]{4,12}$",
                    "title": "Order Number"
                  },
                  "amount": {
                    "type": "string",
                    "maxLength": 16,
                    "title": "Amount",
                    "description": "Decimal amount as a string, e.g. '49.90'. Never a float."
                  },
                  "currency": {
                    "type": "string",
                    "maxLength": 3,
                    "minLength": 3,
                    "title": "Currency",
                    "description": "ISO currency code; must match the order"
                  },
                  "reason": {
                    "type": "string",
                    "enum": [
                      "damaged_on_arrival",
                      "not_delivered",
                      "wrong_item",
                      "duplicate_charge",
                      "other"
                    ],
                    "description": "Why the refund is proposed. One of the listed values exactly — not free text, and not the customer's wording."
                  },
                  "note": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 500
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Note"
                  }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "order_number",
                  "amount",
                  "currency",
                  "reason"
                ],
                "title": "ProposeRefundRequest",
                "description": "The refund proposal body.\n\nAbsent by design: any identity field, any approval field, any destination. The destination is\nderived server-side from the order at execution time — a caller-supplied payout target is how\nrefund fraud works."
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposeRefundResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Not found or not allowed"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
    "/v1/actions/{action_id}": {
      "get": {
        "tags": [
          "actions"
        ],
        "summary": "Get the status of a previously requested action",
        "operationId": "get_action_status",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F-]{36}$",
              "description": "The action identifier returned by propose_refund",
              "title": "Action Id"
            },
            "description": "The action identifier returned by propose_refund"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Not found or not allowed"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionStatusResponse": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "provider_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Reference"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_id",
          "action_type",
          "state",
          "created_at"
        ],
        "title": "ActionStatusResponse"
      },
      "OrderItem": {
        "properties": {
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_amount": {
            "type": "string",
            "title": "Unit Amount"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sku",
          "description",
          "quantity",
          "unit_amount"
        ],
        "title": "OrderItem"
      },
      "OrderResponse": {
        "properties": {
          "order_number": {
            "type": "string",
            "title": "Order Number",
            "description": "The order identifier, e.g. ORD-2001"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Order lifecycle status"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "total_amount": {
            "type": "string",
            "title": "Total Amount",
            "description": "Decimal amount as a string, to avoid float rounding"
          },
          "placed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Placed At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrderItem"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "shipment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Shipment"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "order_number",
          "status",
          "currency",
          "total_amount",
          "placed_at",
          "updated_at"
        ],
        "title": "OrderResponse",
        "description": "The bounded response schema.\n\n`extra=\"forbid\"` means a field the repository adds later cannot reach the caller until someone\nadds it here deliberately — response minimization by construction rather than by review."
      },
      "ProposeRefundResponse": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "state": {
            "type": "string",
            "title": "State",
            "description": "Always PENDING_APPROVAL: no money has moved"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_id",
          "state",
          "expires_at"
        ],
        "title": "ProposeRefundResponse"
      },
      "Shipment": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "carrier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carrier"
          },
          "tracking_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Ref"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "updated_at"
        ],
        "title": "Shipment"
      }
    }
  }
}